+++ skia-bindings/src/gl.cpp
@@ -231,3 +231,13 @@ extern "C" const GrGLInterface* C_GrGLInterfaces_MakeMac() {
 }
 
 #endif
+
+#if defined(SK_BUILD_FOR_WIN)
+
+#include "include/gpu/ganesh/gl/mac/GrGLMakeWinInterface.h"
+
+extern "C" const GrGLInterface* C_GrGLInterfaces_MakeWin() {
+    return GrGLInterfaces::MakeWin().release();
+}
+
+#endif
