+++ skia-safe/src/core.rs
@@ -33,6 +33,7 @@ mod font_mgr;
 pub mod font_parameters;
 pub mod font_style;
 mod font_types;
+mod four_byte_tag;
 pub mod graphics;
 pub mod image;
 pub mod image_filter;
@@ -110,6 +111,7 @@ pub use font_metrics::FontMetrics;
 pub use font_mgr::*;
 pub use font_style::FontStyle;
 pub use font_types::*;
+pub use four_byte_tag::*;
 pub use image::{images, Image};
 pub use image_filter::ImageFilter;
 pub use image_generator::*;
