+++ skia-safe/src/gpu/ganesh.rs
@@ -1,8 +1,17 @@
+pub mod backend_surface;
+pub mod context_options;
+#[cfg(feature = "d3d")]
+pub mod d3d;
+pub mod direct_context;
+pub mod driver_bug_workarounds;
 #[cfg(feature = "gl")]
 pub mod gl;
 pub mod image_ganesh;
 #[cfg(feature = "metal")]
 pub mod mtl;
+pub mod recording_context;
 pub mod surface_ganesh;
+pub mod types;
 #[cfg(feature = "vulkan")]
 pub mod vk;
+pub mod yuva_backend_textures;
