+++ skia-safe/src/gpu/ganesh/gl.rs
@@ -1,5 +1,7 @@
 mod backend_surface;
 mod direct_context;
+pub mod extensions;
+pub mod interface;
 #[cfg(feature = "egl")]
 pub mod make_egl_interface;
 #[cfg(target_os = "ios")]
@@ -8,6 +10,9 @@ pub mod make_ios_interface;
 pub mod make_mac_interface;
 #[cfg(target_arch = "wasm32")]
 pub mod make_web_gl_interface;
+#[cfg(target_os = "windows")]
+pub mod make_win_interface;
+pub mod types;
 
 pub use backend_surface::*;
 pub use direct_context::*;
