+++ skia-safe/src/gpu/ganesh/gl/make_win_interface.rs
@@ -1,0 +1,9 @@
+pub mod interfaces {
+    use skia_bindings as sb;
+
+    use crate::gpu::gl;
+
+    pub fn make_win() -> Option<gl::Interface> {
+        gl::Interface::from_ptr(unsafe { sb::C_GrGLInterfaces_MakeWin() } as *mut _)
+    }
+}
