+++ skia-safe/src/gpu/ganesh/vk.rs
@@ -1,6 +1,7 @@
 mod backend_drawable_info;
 mod vk_backend_surface;
 mod vk_direct_context;
+pub mod vk_types;
 
 pub use backend_drawable_info::*;
 pub use vk_backend_surface::*;
