+++ skia-safe/src/gpu/ganesh/yuva_backend_textures.rs
@@ -1,4 +1,4 @@
-use super::{BackendFormat, BackendTexture, Mipmapped, SurfaceOrigin};
+use crate::gpu::{BackendFormat, BackendTexture, Mipmapped, SurfaceOrigin};
 use crate::{prelude::*, YUVAInfo, YUVColorSpace};
 use skia_bindings::{self as sb, GrYUVABackendTextureInfo, GrYUVABackendTextures};
 use std::{fmt, iter};
